TldLogExporter and EventHeaderLogExporter works, but TldTraceExporter is also impacted since it implements TldExporter
