+++ src/OpenTelemetry.Exporter.Geneva/Internal/EventHeader/EventHeaderLogExporter.cs
@@ -3,10 +3,7 @@
 
 #if NET
 
-using System.Diagnostics;
 using System.Globalization;
-using System.Runtime.CompilerServices;
-using Microsoft.Extensions.Logging;
 using Microsoft.LinuxTracepoints.Provider;
 using OpenTelemetry.Exporter.Geneva.Tld;
 using OpenTelemetry.Exporter.Geneva.Transports;
@@ -18,84 +15,25 @@ namespace OpenTelemetry.Exporter.Geneva.EventHeader;
 internal class EventHeaderLogExporter : EventHeaderExporter, IDisposable
 {
     internal const int StringLengthLimit = (1 << 14) - 1; // 16 * 1024 - 1 = 16383
-    private const int MaxSanitizedEventNameLength = 50;
 
     private static readonly ThreadLocal<EventHeaderDynamicBuilder> EventBuilder = new();
-    private static readonly ThreadLocal<KeyValuePair<string, object>[]> PartCFields = new(); // This is used to temporarily store the PartC fields from tags
-    private static readonly ThreadLocal<List<KeyValuePair<string, object?>>> EnvProperties = new();
     private static readonly Action<LogRecordScope, EventHeaderLogExporter> ProcessScopeForIndividualColumnsAction = OnProcessScopeForIndividualColumns;
 
-    private static readonly string[] LogLevels =
-    [
-        "Trace", "Debug", "Information", "Warning", "Error", "Critical", "None"
-    ];
-
-    private readonly ThreadLocal<SerializationDataForScopes> serializationData = new(); // This is used for Scopes
-    private readonly byte partAFieldsCount = 1; // At least one field: time
-    private readonly bool shouldPassThruTableMappings;
-    private readonly string defaultEventName = "Log";
-    private readonly HashSet<string>? customFields;
-    private readonly Dictionary<string, string>? tableMappings;
     private readonly ValueTuple<byte[], byte[]>? repeatedPartAFields;
-    private readonly ExceptionStackExportMode exceptionStackExportMode;
 
 #pragma warning disable CA2213 // Disposable fields should be disposed: it is managed and disposed in the Dispose method of EventHeaderDynamicProvider
     private readonly EventHeaderDynamicTracepoint logsTracepoint;
 #pragma warning restore CA2213 // Disposable fields should be disposed
 
-    private bool isDisposed;
-
     public EventHeaderLogExporter(GenevaExporterOptions options)
+        : base(options)
     {
         Guard.ThrowIfNull(options);
 
         this.logsTracepoint = UnixUserEventsDataTransport.Instance.RegisterUserEventProviderForLogs();
 
-        this.exceptionStackExportMode = options.ExceptionStackExportMode;
-
-        // TODO: Validate mappings for reserved tablenames etc.
-        if (options.TableNameMappings != null)
-        {
-            var tempTableMappings = new Dictionary<string, string>(options.TableNameMappings.Count, StringComparer.Ordinal);
-            foreach (var kv in options.TableNameMappings)
-            {
-                if (kv.Key == "*")
-                {
-                    if (kv.Value == "*")
-                    {
-                        this.shouldPassThruTableMappings = true;
-                    }
-                    else
-                    {
-                        this.defaultEventName = kv.Value;
-                    }
-                }
-                else
-                {
-                    tempTableMappings[kv.Key] = kv.Value;
-                }
-            }
-
-            this.tableMappings = tempTableMappings;
-        }
-
-        // TODO: Validate custom fields (reserved name? etc).
-        if (options.CustomFields != null)
-        {
-            var customFields = new HashSet<string>(StringComparer.Ordinal);
-            foreach (var name in options.CustomFields)
-            {
-                customFields.Add(name);
-            }
-
-            this.customFields = customFields;
-        }
-
         if (options.PrepopulatedFields != null)
         {
-            var prePopulatedFieldsCount = (byte)(options.PrepopulatedFields.Count - 1); // PrepopulatedFields option has the key ".ver" added to it which is not needed for TLD
-            this.partAFieldsCount += prePopulatedFieldsCount;
-
             var eb = EventBuilder.Value ??= new EventHeaderDynamicBuilder();
 
             eb.Reset("_"); // EventName does not matter here as we only need the serialized key-value pairs
@@ -110,7 +48,7 @@ internal class EventHeaderLogExporter : EventHeaderExporter, IDisposable
                     continue;
                 }
 
-                V40_PART_A_EVENT_HEADER_MAPPING.TryGetValue(key, out var replacementKey);
+                V40_PART_A_TLD_MAPPING.TryGetValue(key, out var replacementKey);
                 var keyToSerialize = replacementKey ?? key;
                 Serialize(eb, keyToSerialize, value);
             }
@@ -142,26 +80,6 @@ internal class EventHeaderLogExporter : EventHeaderExporter, IDisposable
         }
 
         return ExportResult.Failure;
-    }
-
-    public void Dispose()
-    {
-        if (this.isDisposed)
-        {
-            return;
-        }
-
-        try
-        {
-            // DO NOT Dispose eventBuilder, keyValuePairs, and partCFields as they are static
-            this.serializationData?.Dispose();
-        }
-        catch (Exception ex)
-        {
-            ExporterEventSource.Log.ExporterException("EventHeaderLogExporter Dispose failed.", ex);
-        }
-
-        this.isDisposed = true;
     }
 
     internal EventHeaderDynamicBuilder SerializeLogRecord(LogRecord logRecord)
@@ -421,127 +339,15 @@ internal class EventHeaderLogExporter : EventHeaderExporter, IDisposable
         return eb;
     }
 
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static byte GetSeverityNumber(LogLevel logLevel)
-    {
-        // Maps the Ilogger LogLevel to OpenTelemetry logging level.
-        // https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/logs/data-model.md#mapping-of-severitynumber
-        // TODO: for improving perf simply do ((int)loglevel * 4) + 1
-        // or ((int)logLevel << 2) + 1
-        return logLevel switch
-        {
-            LogLevel.Trace => 1,
-            LogLevel.Debug => 5,
-            LogLevel.Information => 9,
-            LogLevel.Warning => 13,
-            LogLevel.Error => 17,
-            LogLevel.Critical => 21,
-
-            // we reach default only for LogLevel.None
-            // but that is filtered out anyway.
-            // should we throw here then?
-            LogLevel.None => 1,
-            _ => 1,
-        };
-    }
-
-    // This method would map the logger category to a table name which only contains alphanumeric values with the following additions:
-    // Any character that is not allowed will be removed.
-    // If the resulting string is longer than 50 characters, only the first 50 characters will be taken.
-    // If the first character in the resulting string is a lower-case alphabet, it will be converted to the corresponding upper-case.
-    // If the resulting string still does not comply with Rule, the category name will not be serialized.
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static string GetSanitizedCategoryName(string categoryName)
-    {
-        var validNameLength = 0;
-        Span<char> result = stackalloc char[MaxSanitizedEventNameLength];
-
-        // Special treatment for the first character.
-        var firstChar = categoryName[0];
-        if (firstChar is >= 'A' and <= 'Z')
-        {
-            result[0] = firstChar;
-            ++validNameLength;
-        }
-        else if (firstChar is >= 'a' and <= 'z')
-        {
-            // If the first character in the resulting string is a lower-case alphabet,
-            // it will be converted to the corresponding upper-case.
-            result[0] = (char)(firstChar - 32);
-            ++validNameLength;
-        }
-        else
-        {
-            // Not a valid name.
-            return string.Empty;
-        }
-
-        for (var i = 1; i < categoryName.Length; i++)
-        {
-            if (validNameLength == MaxSanitizedEventNameLength)
-            {
-                break;
-            }
-
-            var cur = categoryName[i];
-            if (cur is (>= 'a' and <= 'z') or (>= 'A' and <= 'Z') or (>= '0' and <= '9'))
-            {
-                result[validNameLength] = cur;
-                ++validNameLength;
-            }
-        }
-
-        return result.Slice(0, validNameLength).ToString();
-    }
+    protected override void Dispose(bool disposing) => base.Dispose(disposing);
 
     private static void OnProcessScopeForIndividualColumns(LogRecordScope scope, EventHeaderLogExporter state)
-    {
-        Debug.Assert(state.serializationData.Value != null, "state.serializationData was null");
-        Debug.Assert(PartCFields.Value != null, "PartCFields.Value was null");
-
-        var stateData = state.serializationData.Value!;
-        var customFields = state.customFields;
-        var kvpArrayForPartCFields = PartCFields.Value!;
-
-        List<KeyValuePair<string, object?>>? envPropertiesList = null;
-
-        foreach (var scopeItem in scope)
-        {
-            if (string.IsNullOrEmpty(scopeItem.Key) || scopeItem.Key == "{OriginalFormat}")
-            {
-                continue;
-            }
-
-            if (customFields == null || customFields.Contains(scopeItem.Key))
-            {
-                if (scopeItem.Value != null)
-                {
-                    kvpArrayForPartCFields[stateData.PartCFieldsCountFromState] = new(scopeItem.Key, scopeItem.Value);
-                    stateData.PartCFieldsCountFromState++;
-                }
-            }
-            else
-            {
-                if (stateData.HasEnvProperties == 0)
-                {
-                    stateData.HasEnvProperties = 1;
-
-                    envPropertiesList = EnvProperties.Value ??= [];
-
-                    envPropertiesList.Clear();
-                }
-
-                // TODO: This could lead to unbounded memory usage.
-                envPropertiesList!.Add(new(scopeItem.Key, scopeItem.Value));
-            }
-        }
-    }
-
-    private sealed class SerializationDataForScopes
-    {
-        public byte HasEnvProperties;
-        public byte PartCFieldsCountFromState;
-    }
+        => OnProcessScopeForIndividualColumns(
+            scope,
+            state.serializationData.Value!,
+            state.customFields,
+            PartCFields.Value!,
+            EnvProperties.Value);
 }
 
 #endif
